Fixed a few compile errors.

+++ src/main/scala/li/cil/oc/util/Audio.scala
@@ -147,8 +147,8 @@ object Audio {
           checkALError()
 
           AL10.alSource3f(source, AL10.AL_POSITION, x, y, z)
-          AL10.alSourcef(source, AL10.AL_REFERENCE_DISTANCE, maxDistance)
-          AL10.alSourcef(source, AL10.AL_MAX_DISTANCE, maxDistance)
+          AL10.alSourcef(source, AL10.AL_REFERENCE_DISTANCE, maxDistance.toFloat)
+          AL10.alSourcef(source, AL10.AL_MAX_DISTANCE, maxDistance.toFloat)
           AL10.alSourcef(source, AL10.AL_GAIN, gain * 0.3f)
           checkALError()
